requirements.txt
# Pinned to avoid build-from-source on Render with Python 3.12 (runtime.txt provided).
pandas==2.2.2
requests>=2.31.0
python-slugify>=8.0.4
tenacity>=8.2.3
gspread>=6.0.0
google-auth>=2.28.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
